Added food stuff
All the food stuff and user stuff is now prepared and should be ready for migration to database. Since i dont know how to handle multiple context's in the same assembly i will probably either horribly mess it up or have to change something, so this commit acts as a backup.