Reworked foods and fixed many-many
Most notably replaced new() for list objects with null! as i thought props were regarded fields by VS for this reason but apparently not. Does not affect database however so left as is. Defined relations in modelbuilder too, with explicitly defined delete behaviors although they are all still default behaviors. What remains in foods is to create repos and seed.