SQL models and relations seem to be working
I tried a "ghetto-seeding" in mainwindow and found out how to create new entries and their relations worked. As for the data layer the only thing remaining (that i know of so far) is to flesh out the repos with more features (probably as i develop UI layer) and most importantly to create a REAL seeding method, since i couldnt for the life of me get it to work in OnModelCreating.